refactor the code relates #104
