+++ server/database/queries/wishlist/index.ts
@@ -1,4 +1,5 @@
 import addToWishListQuery from './addToWishList';
 import deleteFromWishListQuery from './deleteFromWishList';
+import getAllWishlistItemsQuery from './getAllWishListItemsQuery';
 
-export { addToWishListQuery, deleteFromWishListQuery };
+export { addToWishListQuery, deleteFromWishListQuery, getAllWishlistItemsQuery };
